Moved some functions out to a util
