+++ Makefile
@@ -2,7 +2,7 @@ CC=gcc
 CFLAGS=-g -Wall -Wextra -Wpedantic
 LDFLAGS=-lm -lcurl -ljansson -lSDL2 -lSDL2_mixer
 DEPS=
-OBJ = lib/main.c lib/http.c lib/words.c lib/sound.c lib/cli.c
+OBJ = lib/main.c lib/http.c lib/words.c lib/sound.c lib/cli.c lib/util.c
 
 %.o: %.c $(DEPS)
 	$(CC) -c -o $@ $< $(CFLAGS)
