+++ lib/util.h
@@ -1,0 +1,9 @@
+#include <stdbool.h>
+char upper(char c);
+bool is_alphabetic(char c);
+bool str_contains(char* str, char c);
+
+bool bit_set(unsigned bm, int pos);
+unsigned set_bit(unsigned bm, int pos);
+
+unsigned letter_positions(char* str, char c);
